+++ src/main/kotlin/com/alloon/alloonserver/domain/mission/MissionHashtag.kt
@@ -2,7 +2,6 @@ package com.alloon.alloonserver.domain.mission
 
 import com.alloon.alloonserver.domain.base.BasePermanentEntity
 import jakarta.persistence.*
-import java.math.BigInteger
 
 @Entity
 class MissionHashtag(
@@ -10,7 +9,7 @@ class MissionHashtag(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "mission_hashtag_id")
-    val id: BigInteger? = null,
+    val id: Long? = null,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "mission_id", nullable = false)
