+++ src/main/kotlin/com/alloon/alloonserver/domain/user/ContactRepository.kt
@@ -1,9 +1,8 @@
 package com.alloon.alloonserver.domain.user
 
 import org.springframework.data.jpa.repository.JpaRepository
-import java.math.BigInteger
 
-interface ContactRepository : JpaRepository<Contact, BigInteger> {
+interface ContactRepository : JpaRepository<Contact, Long> {
 
     fun findByEmail(email: String): Contact?
 }
