+++ src/main/kotlin/com/alloon/alloonserver/domain/user/UserRepository.kt
@@ -1,9 +1,8 @@
 package com.alloon.alloonserver.domain.user
 
 import org.springframework.data.jpa.repository.JpaRepository
-import java.math.BigInteger
 
-interface UserRepository : JpaRepository<User, BigInteger> {
+interface UserRepository : JpaRepository<User, Long> {
 
     fun existsByContact(contact: Contact): Boolean
 }
