+++ src/main/kotlin/com/alloon/alloonserver/domain/user/UserRole.kt
@@ -2,8 +2,6 @@ package com.alloon.alloonserver.domain.user
 
 import com.alloon.alloonserver.domain.base.BaseEntity
 import jakarta.persistence.*
-import java.math.BigInteger
-import java.time.LocalDateTime
 
 @Entity
 class UserRole(
@@ -11,7 +9,7 @@ class UserRole(
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_role_id")
-    val id: BigInteger? = null,
+    val id: Long? = null,
 
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "user_id", nullable = false)
